Time: 0 ms (100.00%) | Memory: 25.7 MB (64.70%) - LeetSync

+++ 3471-minimum-average-of-smallest-and-largest-elements/minimum-average-of-smallest-and-largest-elements.cpp
@@ -1,8 +1,7 @@
 class Solution {
 public:
     double minimumAverage(vector<int>& nums) {
-        
-        double res=INT_MAX;
+        double res=DBL_MAX;
         sort(nums.begin(),nums.end());
         int i=0;
         int j=nums.size()-1;
@@ -14,7 +13,6 @@ public:
              i++;
              j--;
         }
-
         return res;
     }
 };
